add helpers file no more isWord leak, but json5.js has a require so it can't be used in a browser directly browserify will likely need to be implemented first
