+++ lib/json5.js
@@ -4,6 +4,8 @@
 // This file is based directly off of Douglas Crockford's json_parse.js:
 // https://github.com/douglascrockford/JSON-js/blob/master/json_parse.js
 
+var helpers = require('./helpers');
+
 var JSON5 = (typeof exports === 'object' ? exports : {});
 
 JSON5.parse = (function () {
@@ -547,39 +549,6 @@ JSON5.stringify = function (obj, replacer, space) {
         }
     };
 
-    function isWordChar(char) {
-        return (char >= 'a' && char <= 'z') ||
-            (char >= 'A' && char <= 'Z') ||
-            (char >= '0' && char <= '9') ||
-            char === '_' || char === '$';
-    }
-
-    function isWordStart(char) {
-        return (char >= 'a' && char <= 'z') ||
-            (char >= 'A' && char <= 'Z') ||
-            char === '_' || char === '$';
-    }
-
-    function isWord(key) {
-        if (typeof key !== 'string') {
-            return false;
-        }
-        if (!isWordStart(key[0])) {
-            return false;
-        }
-        var i = 1, length = key.length;
-        while (i < length) {
-            if (!isWordChar(key[i])) {
-                return false;
-            }
-            i++;
-        }
-        return true;
-    }
-
-    // export for use in tests
-    JSON5.isWord = isWord;
-
     // polyfills
     function isArray(obj) {
         if (Array.isArray) {
@@ -720,7 +689,7 @@ JSON5.stringify = function (obj, replacer, space) {
                             if (typeof value !== "undefined" && value !== null) {
                                 buffer += makeIndent(indentStr, objStack.length);
                                 nonEmpty = true;
-                                key = isWord(prop) ? prop : escapeString(prop);
+                                var key = helpers.isWord(prop) ? prop : escapeString(prop);
                                 buffer += key + ":" + (indentStr ? ' ' : '') + value + ",";
                             }
                         }
